--- spec/fixtures/postcode_error_conversion_payload.json
@@ -1,4 +1,0 @@
-{
-  "status": 404,
-  "error": "Invalid postcode"
-}
